requirements.txt
customtkinter>=5.2.0
pandas>=2.0.0
openpyxl>=3.1.0
python-dateutil>=2.8.2
Pillow>=10.0.0  # Required by customtkinter
numpy>=1.24.0  # Required by pandas 